requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
brevitas==0.10.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
concrete-ml==1.9.0
concrete-ml-extensions==0.1.9
concrete-python==2.10.0
contourpy==1.3.2
cycler==0.12.1
dependencies==2.0.1
dill==0.4.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.3
huggingface-hub==0.32.6
humanfriendly==10.0
hummingbird-ml==0.4.11
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonpickle==4.1.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.42.1
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxconverter-common==1.13.0
onnxmltools==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.18.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==11.2.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.5.0
scipy==1.15.3
six==1.17.0
skl2onnx==1.19.1
skops==0.5.0
skorch==0.11.0
smmap==5.0.2
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.3.1
tornado==6.5.1
tqdm==4.67.1
triton==2.3.1
typing_extensions==4.14.0
tzdata==2025.2
unfoldNd==0.2.3
urllib3==2.4.0
watchdog==6.0.0
xgboost==1.6.2
z3-solver==4.13.0.0
